Update V-HACD, Enable warnings as errors on MSVC (#102)
* Update v-hacd, enable warnings as errors on MSVC

* Re-comment v-hacd cmake messages

* Revert some changes

* Another revert

* GCC fix

* More GCC warnings

+++ src/core/systems/AnimationSystem.cc
@@ -5,7 +5,7 @@
 #include "ecs/EcsImpl.hh"
 
 namespace sp {
-    bool AnimationSystem::Frame(float dtSinceLastFrame) {
+    bool AnimationSystem::Frame(double dtSinceLastFrame) {
         auto lock = ecs::World.StartTransaction<ecs::Write<ecs::Animation, ecs::Transform>>();
         for (auto ent : lock.EntitiesWith<ecs::Animation>()) {
             if (!ent.Has<ecs::Animation, ecs::Transform>(lock)) continue;
